requirements.txt
numpy==1.21.4
pandas==1.3.4
pandas-ta==0.3.14b

ccxt==1.60.73
# Pin cryptography for now due to rust build errors with piwheels
cryptography==35.0.0
aiohttp==3.7.4.post0
SQLAlchemy==1.4.26
python-telegram-bot==13.7
arrow==1.2.1
cachetools==4.2.2
requests==2.26.0
urllib3==1.26.7
jsonschema==4.2.1
TA-Lib==0.4.21
technical==1.3.0
tabulate==0.8.9
pycoingecko==2.2.0
jinja2==3.0.2
tables==3.6.1
blosc==1.10.6

# find first, C search in arrays
py_find_1st==1.1.5

# Load ticker files 30% faster
python-rapidjson==1.5

# Notify systemd
sdnotify==0.3.2

# API Server
fastapi==0.70.0
uvicorn==0.15.0
pyjwt==2.3.0
aiofiles==0.7.0
psutil==5.8.0

# Support for colorized terminal output
colorama==0.4.4
# Building config files interactively
questionary==1.10.0
prompt-toolkit==3.0.22
# Extensions to datetime library
python-dateutil==2.8.2

# MQTT
paho-mqtt

# NFI
pandas-ta